Added pagination; code clean up
