+++ angular/src/app/search/filters/filters.component.html
@@ -1,0 +1,130 @@
+<div [ngStyle]="{'height': comheight, 'width': filterWidthNum+'px'}">
+    <div class="normalFilter collapsed in width" [hidden]="!isActive" id="activeFilter">
+        <div style="padding: 0px;">
+
+            <!-- Filter header -->
+            <div id="filterHeader">
+                <a id="filtersLabel">Filters </a> <a
+                    class="btn-sm collapsed in" (click)="setColumnWidth()" data-toggle="collapse"
+                    style="border-color: #025277; color: #FFF;padding: 0em;border: 0em"
+                    data-target="#demo"><i class="faa faa-angle-double-left"
+                        style="zoom: 100%;color: #fff;cursor: pointer;"></i></a>
+                <a href="#" id="clear-all"
+                    (click)="$event.preventDefault();clearFilters()">
+                    <i style="vertical-align: middle;zoom: 70%;margin-right: .3em;" class="faa faa-remove">
+                    </i>Clear All</a>
+            </div>
+
+            <!-- Filter body -->
+            <!-- Display spinner while waiting for search result -->
+            <div class="spinner" *ngIf="searching">
+                <i class="faa faa-spinner faa-spin faa-stack-2x" style="color:#1E6BA1; margin: 1em auto;" aria-hidden="true"></i>
+            </div>
+
+            <div style="padding-left: .5em;">
+                <div style="border-bottom:.5px solid #e6e6e6;width: 100%;">
+                    <!-- Resource type -->
+                    <p-tree [value]="resourceTypeTree" selectionMode="checkbox" [style]="ResourceTypeStyle"
+                        [(selection)]="selectedResourceTypeNode" 
+                        (onNodeUnselect)="filterResults($event,'')"
+                        (onNodeSelect)="filterResults($event,'')">
+                        <ng-template let-node pTemplate="default">
+                            <span> {{node.label.split("-")[0]}}&nbsp;</span>
+                            <span class="w3-badge badge1" style="background-color: #1471AE;"
+                                *ngIf="node.label.split('-')[1] !=''">
+                                {{node.label.split("-")[1]}}</span>
+                        </ng-template>
+                    </p-tree>
+                </div>
+
+                <!-- Research topics -->
+                <div id="research-topic" >
+                    <p-tree header="test" id="themes" [value]="themesTree" selectionMode="checkbox" [style]="getResearchTopicStyle()" [(selection)]="selectedThemesNode"
+                        (onNodeUnselect)="filterResults($event,'')"
+                        (onNodeSelect)="filterResults($event,'')" 
+                        (onNodeExpand)="nodeExpanded = true"
+                        (onNodeCollapse)="nodeExpanded = false">
+                        <ng-template let-node pTemplate="default">
+                            <span class="invisible-scrollbar"> {{node.label.split("-")[0]}}&nbsp;</span>
+                            <span class="w3-badge badge1 invisible-scrollbar"
+                                style="background-color: #1471AE" *ngIf="node.label.split('-')[1] !=''">
+                                {{node.label.split("-")[1]}}</span>
+                        </ng-template>
+                    </p-tree>
+                    <div style="background-color: #F8F9F9"
+                        *ngIf="showMoreResearchTopics == false && nodeExpanded == true && showMoreLink == true">
+                        <a href="#" (click)="$event.preventDefault();showMoreResearchTopics = true"
+                            style="background-color: #F8F9F9;color: #1f496e;padding-left: 50px;"
+                            *ngIf="showMoreResearchTopics == false && nodeExpanded == true && showMoreLink == true">Show
+                            More...</a>
+                    </div>
+                    <div style="background-color: #F8F9F9"
+                        *ngIf="showMoreResearchTopics == true && nodeExpanded == true && showMoreLink == true">
+                        <a href="#" (click)="$event.preventDefault();showMoreResearchTopics = false;"
+                            style="background-color: #F8F9F9;color: #1f496e;padding-left: 50px;">Show
+                            Less</a>
+                    </div>
+                </div>
+
+                <!-- Components -->
+                <div style="border-bottom:.5px solid #e6e6e6;width: 100%;">
+                    <p-tree [value]="componentsTree" selectionMode="checkbox" [style]="{'width':'auto','padding-top': '.5em','padding-right': '.5em',
+            'padding-bottom': '.5em','background-color': '#F8F9F9'}"
+                        [(selection)]="selectedComponentsNode" (onNodeUnselect)="filterResults($event,'')"
+                        (onNodeSelect)="filterResults($event,'')">
+                        <ng-template let-node pTemplate="default">
+                            <span> {{node.label.split("-")[0]}}&nbsp;</span>
+                            <span class="w3-badge badge1" style="background-color: #1471AE"
+                                *ngIf="node.label.split('-')[1] !=''">
+                                {{node.label.split("-")[1]}}</span>
+                        </ng-template>
+                    </p-tree>
+                </div>
+
+                <!-- Authors and Contributors -->
+                <div style="padding: 0px;">
+                    <div style="background-color:#F8F9F9;width: 100%;padding-left: 1em;padding-top:1em;">
+                        <label><b>Authors and Contributors</b></label>
+                    </div>
+                    <div style="background-color: #F8F9F9;padding-left:15px;padding-right:1.0em;">
+                        <label for="suggestedauthor" class="hideLabel">Search Input</label>
+                        <p-autoComplete [immutable]="true" [(ngModel)]="selectedAuthor"
+                            inputId="suggestedauthor" [suggestions]="suggestedAuthors"
+                            (completeMethod)="filterAuthors($event)" [multiple]="true"
+                            (onSelect)="filterResults($event,'')"
+                            (onUnselect)="filterResults($event,'')" [minLength]="1"
+                            [maxlength]="30" [style]="filterStyle">
+                        </p-autoComplete>
+                    </div>
+                </div>
+
+                <!-- Keyword -->
+                <div style="padding: 0px;">
+                    <div
+                        style="border-top:.5px solid #e6e6e6;margin-top:1em;background-color:#F8F9F9;width: 100%;padding-left: 1em;padding-top:1em;">
+                        <label><b>Keyword</b></label>
+                    </div>
+                    <div style="background-color: #F8F9F9;padding-left:15px;padding-right:1.0em;">
+                        <label for="keyword" class="hideLabel">Keyword</label>
+                        <p-autoComplete inputId="keyword" [(ngModel)]="selectedKeywords"
+                            [suggestions]="suggestedKeywords" (completeMethod)="filterKeywords($event)"
+                            [multiple]="true" (onSelect)="filterResults($event,'')"
+                            (onUnselect)="filterResults($event,'')" [minLength]="1"
+                            [maxlength]="30" [style]="filterStyle">
+                        </p-autoComplete>
+                    </div>
+                </div>
+            </div>
+        </div>
+    </div>
+
+    <div *ngIf="!searching" [hidden]="isActive" class="collapsedFilter"
+         [ngStyle]="{'height': comheight}"> <br>
+        
+        <!-- Collapsed filters -->
+        <div (click)="setColumnWidth()">
+            <i class="faa faa-angle-double-right" style="color: #fff;zoom: 130%"></i>
+            <div class="rotate rotatedFilter"> Filters </div>
+        </div>
+    </div>
+</div>
