+++ angular/src/app/search/pagination/pagination.component.html
@@ -1,0 +1,25 @@
+<div>
+    <ul *ngIf="pager.pages && pager.pages.length" class="pagination">
+        <li [ngClass]="{disabled:pager.currentPage === 1}">
+            <a (click)="setPage(pager.currentPage - 1)">Previous</a>
+        </li>
+        <li [ngClass]="{active:pager.currentPage === 1}">
+            <a (click)="setPage(1)">1</a>
+        </li>
+        <li [ngClass]="{disabled:true,hide:pager.currentPage < 4 || pager.totalPages < 6}">
+            <a (click)="setPage(pager.currentPage + 1)">...</a>
+        </li>
+        <li *ngFor="let page of pager.pages" [ngClass]="{active:pager.currentPage === page, hide:pager.totalPages<3}">
+            <a (click)="setPage(page)">{{page}}</a>
+        </li>
+        <li [ngClass]="{disabled:true,hide:pager.currentPage > pager.totalPages-3 || pager.totalPages < 6}">
+            <a (click)="setPage(pager.currentPage + 1)">...</a>
+        </li>
+        <li [ngClass]="{active:pager.currentPage === pager.totalPages,hide:pager.totalPages==1}">
+            <a (click)="setPage(pager.totalPages)">{{pager.totalPages}}</a>
+        </li>
+        <li [ngClass]="{disabled:pager.currentPage === pager.totalPages}">
+            <a (click)="setPage(pager.currentPage + 1)">Next</a>
+        </li>
+    </ul>
+</div>
