+++ angular/src/app/search/results/results.component.css
@@ -1,0 +1,21 @@
+.title {
+  color:#1471AE;
+  font-size: 15px;
+}
+
+.flex-container {
+    width: 100%;
+    /* height: 2.8em; */
+    display: flex;
+    align-items: stretch;
+    /* background-color: #f1f1f1; */
+}
+
+.flex-grow1 {
+    flex-grow: 1;
+}
+
+.full-width {
+    width: 100%;
+    text-align: center;
+}
