+++ angular/src/app/search/search.component.css
@@ -397,3 +397,4 @@ background-cyanblue {
     text-align: left;
     background-color: rgb(255, 226, 171);
 }
+
