+++ angular/src/app/shared/search-service/search-service.service.mock.ts
@@ -14,59 +14,89 @@ import { Router, NavigationExtras } from '@angular/router';
 import { SDPQuery } from '../search-query/query';
 
 @Injectable({
-  providedIn: 'root'
+    providedIn: 'root'
 })
 
 export class MockSearchService implements SearchService{
-  confValues: Config;
-  private RMMAPIURL: string;
-
-  /**
-   * Creates a new SearchService with the injected Http.
-   * @param {HttpClient} http - The injected Http.
-   * @constructor
-   */
-  constructor(private http: HttpClient,
-    private router: Router,
-    private appConfig: AppConfig) {
-    this.confValues = this.appConfig.getConfig();
-    this.RMMAPIURL = this.confValues.RMMAPI;
-  }
-
-  /**
-   * Returns an Observable for the HTTP GET request for the JSON resource.
-   * @return {string[]} The Observable for the HTTP request.
-   */
-
-  searchPhrase(query: SDPQuery, searchTaxonomyKey: string, queryAdvSearch: string): Observable<any> {
-    var mockSearchResult = require('../../../assets/sample01.json');
-    return of(mockSearchResult);
-  }
+    confValues: Config;
+    private RMMAPIURL: string;
+    filterString = new BehaviorSubject<string>('');
 
     /**
-   * Returns an Observable for the HTTP GET request for the JSON resource.
-   * @return {string[]} The Observable for the HTTP request.
-   */
-  searchPhraseTest(searchValue: string, searchTaxonomyKey: string, queryAdvSearch: string): Observable<any> {
-    if ((queryAdvSearch === 'yes' && (!(_.includes(searchValue, 'searchphrase'))))) {
-      return this.http.get(this.RMMAPIURL + 'records?' + searchValue);
-    } else {
-      let params = new HttpParams();
-      params.set('searchphrase', searchValue);
-      params.set('topic.tag', searchTaxonomyKey);
+     * Creates a new SearchService with the injected Http.
+     * @param {HttpClient} http - The injected Http.
+     * @constructor
+     */
+    constructor(private http: HttpClient,
+                private router: Router,
+                private appConfig: AppConfig) {
 
-      return this.http.get(this.RMMAPIURL + 'records?',
-        {
-          params: params
-
-        });
+        this.confValues = this.appConfig.getConfig();
+        this.RMMAPIURL = this.confValues.RMMAPI;
     }
-  }
+
+    /**
+     * Returns an Observable for the HTTP GET request for the JSON resource.
+     * @return {string[]} The Observable for the HTTP request.
+     */
+    searchPhrase(
+                query: SDPQuery, 
+                searchTaxonomyKey: string, 
+                queryAdvSearch: string, 
+                page?: number, 
+                pageSize?: number, 
+                sortOrder?:string,
+                filter?:string): Observable<any> {
+
+        var mockSearchResult = require('../../../assets/sample01.json');
+        return of(mockSearchResult);
+    }
+
+    /**
+     * Watch the filter string
+     */
+    watchFilterString(subscriber){
+        return this.filterString.subscribe(subscriber);
+    }
+
+    /**
+     * Set the filter string
+     **/
+    setFilterString(filterString: string) {
+        this.filterString.next(filterString);
+    }
+
+    //   simpleSearch(page: number, pageSize: number, sortOrder:string): Observable<any> {
+    //     var mockSearchResult = require('../../../assets/sample01.json');
+    //     return of(mockSearchResult);
+    //   }
+    /**
+     * Returns an Observable for the HTTP GET request for the JSON resource.
+     * @return {string[]} The Observable for the HTTP request.
+     */
+    searchPhraseTest(searchValue: string, searchTaxonomyKey: string, queryAdvSearch: string): Observable<any> {
+        if ((queryAdvSearch === 'yes' && (!(_.includes(searchValue, 'searchphrase'))))) {
+
+            return this.http.get(this.RMMAPIURL + 'records?' + searchValue);
+
+        } else {
+            let params = new HttpParams();
+            params.set('searchphrase', searchValue);
+            params.set('topic.tag', searchTaxonomyKey);
+
+            return this.http.get(this.RMMAPIURL + 'records?',
+            {
+                params: params
+            });
+        }
+    }
 
     /**
      * Set/watch the search value in the search text box
      */
-    private _remoteQueryValue : BehaviorSubject<object> = new BehaviorSubject<object>({queryString: '', searchTaxonomyKey: '', queryAdvSearch: 'yes'});
+    private _remoteQueryValue : BehaviorSubject<object> = new BehaviorSubject<object>({
+        queryString: '', searchTaxonomyKey: '', queryAdvSearch: 'yes'});
+
     public _watchQueryValue(subscriber){
         this._remoteQueryValue.subscribe(subscriber);
     }
@@ -79,11 +109,13 @@ export class MockSearchService implements SearchService{
      * Start search
      */
     public search(searchValue: string, url?: string) : void {
+
         let params: NavigationExtras = {
             queryParams: {
                 'q': searchValue
             }
         };
+
         this.router.navigate(['/search'], params);
     }
 }
