+++ angular/src/app/shared/search-service/search-service.service.real.ts
@@ -17,78 +17,129 @@ import { SDPQuery } from '../search-query/query';
  * This class provides the Search service with methods to search for records from tha rmm.
  */
 @Injectable({
-  providedIn: 'root'
+    providedIn: 'root'
 })
 
 export class RealSearchService implements SearchService{
-  confValues: Config;
-  private RMMAPIURL: string;
-  operators = {
-      'AND': 'logicalOp=AND',
-      'OR': 'logicalOp=OR',
-      'NOT': 'logicalOp=NOT'
-  }
-
-  /**
-   * Creates a new SearchService with the injected Http.
-   * @param {HttpClient} http - The injected Http.
-   * @constructor
-   */
-  constructor(private http: HttpClient,
-    private router: Router,
-    private appConfig: AppConfig) {
-    this.confValues = this.appConfig.getConfig();
-    this.RMMAPIURL = this.confValues.RMMAPI;
-  }
-
-  /**
-   * Returns an Observable for the HTTP GET request for the JSON resource.
-   * @return {string[]} The Observable for the HTTP request.
-   */
-  searchPhrase(query: SDPQuery, searchTaxonomyKey: string, queryAdvSearch?: string): Observable<any> {
-    let searchPhraseValue = '';
-    let finalKeyValueStr = '';
-
-    if(query.freeText != null && query.freeText != undefined && query.freeText.trim()!= "")
-        searchPhraseValue = 'searchphrase=' + query.freeText.trim();
-
-    // Processing rows
-    for (let i = 0; i < query.queryRows.length; i++) {
-        if (typeof query.queryRows[i].operator === 'undefined') {
-            query.queryRows[i].operator = 'AND';
-        }
-
-        //Skip operator for the first row
-        if(i > 0){
-            if(query.queryRows[i].operator.trim() == "AND")
-                finalKeyValueStr += '&';
-            else
-                finalKeyValueStr += '&' + this.operators[query.queryRows[i].operator] + '&';
-        }
-
-        //If user didn't provide search value, ignore the row
-        if(!this.isEmpty(query.queryRows[i].fieldText) && !this.isEmpty(query.queryRows[i].fieldValue)){
-            if(finalKeyValueStr[finalKeyValueStr.length-1] != "&")
-                finalKeyValueStr = finalKeyValueStr.trim() + " ";
-
-            finalKeyValueStr += query.queryRows[i].fieldValue + '=' + query.queryRows[i].fieldText.replace(/"/g, ''); 
-        }
+    confValues: Config;
+    private RMMAPIURL: string;
+    operators = {
+        'AND': 'logicalOp=AND',
+        'OR': 'logicalOp=OR',
+        'NOT': 'logicalOp=NOT'
     }
 
-    let keyString: string = '';
-    if(searchTaxonomyKey){
-        keyString = '&topic.tag=' + searchTaxonomyKey;
+    filterString = new BehaviorSubject<string>('');
+
+    /**
+     * Creates a new SearchService with the injected Http.
+     * @param {HttpClient} http - The injected Http.
+     * @constructor
+     */
+    constructor(private http: HttpClient,
+        private router: Router,
+        private appConfig: AppConfig) {
+        this.confValues = this.appConfig.getConfig();
+        this.RMMAPIURL = this.confValues.RMMAPI;
     }
 
-    let url = this.RMMAPIURL + 'records?' + searchPhraseValue.trim();
-    
-    if(searchPhraseValue.trim()!="" && finalKeyValueStr.trim() != "")
-        url += '&'
+    /**
+     * Watch the filter string
+     */
+    watchFilterString(subscriber){
+        return this.filterString.subscribe(subscriber);
+    }
+
+    /**
+     * Set the filter string
+     **/
+    setFilterString(filterString: string) {
+        this.filterString.next(filterString);
+    }
+
+    /**
+     * Returns an Observable for the HTTP GET request for the JSON resource.
+     * @return {string[]} The Observable for the HTTP request.
+     */
+    searchPhrase(query: SDPQuery, searchTaxonomyKey: string, queryAdvSearch?: string, page?: number, pageSize?: number, sortOrder?:string, filter?:string): Observable<any> {
+
+        let searchPhraseValue = '';
+        let finalKeyValueStr = '';
+
+        if(query.freeText != null && query.freeText != undefined && query.freeText.trim()!= "")
+            searchPhraseValue = 'searchphrase=' + query.freeText.trim();
+
+        // Processing rows
+        for (let i = 0; i < query.queryRows.length; i++) {
+            if (typeof query.queryRows[i].operator === 'undefined') {
+                query.queryRows[i].operator = 'AND';
+            }
+
+            //Skip operator for the first row
+            if(i > 0){
+                if(query.queryRows[i].operator.trim() == "AND")
+                    finalKeyValueStr += '&';
+                else
+                    finalKeyValueStr += '&' + this.operators[query.queryRows[i].operator] + '&';
+            }
+
+            //If user didn't provide search value, ignore the row
+            if(!this.isEmpty(query.queryRows[i].fieldText) && !this.isEmpty(query.queryRows[i].fieldValue)){
+                if(finalKeyValueStr[finalKeyValueStr.length-1] != "&")
+                    finalKeyValueStr = finalKeyValueStr.trim() + " ";
+
+                finalKeyValueStr += query.queryRows[i].fieldValue + '=' + query.queryRows[i].fieldText.replace(/"/g, ''); 
+            }
+        }
+
+        let keyString: string = '';
+        if(searchTaxonomyKey){
+            keyString = '&topic.tag=' + searchTaxonomyKey;
+        }
+
+        let url = this.RMMAPIURL + 'records?include=ediid,description,title,keyword,topic.tag,contactPoint,components,@type,doi,landingPage&exclude=_id';
         
-    url += finalKeyValueStr.trim() + keyString.trim();
-    // console.log('url', url);
-    return this.http.get(url);
-  }
+        if(searchPhraseValue)
+            url += "&" + searchPhraseValue.trim();
+
+        if(page){
+            url += '&page=' + page;
+            if(pageSize)
+                url += '&size=' + pageSize;
+            else //default
+                url += '&size=10';
+        }
+        
+        if(sortOrder){
+            url += '&sort.asc=' + sortOrder;
+        }
+
+        if(finalKeyValueStr.trim() != ""){
+            url += '&' + finalKeyValueStr.trim();
+        }
+
+        if(keyString)
+            url += "&" + keyString.trim();
+            
+        if(filter){
+            url += "&" + filter.trim();
+        }
+
+        // console.log('url', url);
+        return this.http.get(url);
+    }
+
+    /**
+     * Returns an Observable for the HTTP GET request for the JSON resource.
+     * @return {string[]} The Observable for the HTTP request.
+     */
+    // simpleSearch(page: number, pageSize: number, sortOrder:string): Observable<any> {
+    //         let url = this.RMMAPIURL + 'records?include=ediid,description,title,keyword,topic.tag,contactPoint,components,@type,doi,landingPage&page=' + page + '&size=' + pageSize + '&sort.asc=' + sortOrder + '&exclude=_id';
+
+    //         console.log("simple http request...", url);
+    //         // console.log('url', url);
+    //         return this.http.get(url);
+    // }
 
     /**
      * Check if a string object is empty
@@ -98,7 +149,7 @@ export class RealSearchService implements SearchService{
         return stringValue == null || stringValue == undefined || stringValue.trim() == '';
     }
 
-      /**
+    /**
      * Convert the text from search text box into url parameter. For example,
      * convert "Research Topic=water" to "topic.tag%3Dwater"
      * convert "Research Topic=water OR Research Topic=fire" to "topic.tag%3Dwater&logicalOp%3DOR=&topic.tag%3Dfire"
@@ -142,56 +193,56 @@ export class RealSearchService implements SearchService{
         return searchString;
     }
 
-  /**
-   * Returns an Observable for the HTTP GET request for the JSON resource.
-   * @return {string[]} The Observable for the HTTP request.
-   */
-  searchPhraseTest(searchValue: string, searchTaxonomyKey: string, queryAdvSearch: string): Observable<any> {
-    if ((queryAdvSearch === 'yes' && (!(_.includes(searchValue, 'searchphrase'))))) {
-      return this.http.get(this.RMMAPIURL + 'records?' + searchValue);
-    } else {
-      let params = new HttpParams();
-      params.set('searchphrase', searchValue);
-      params.set('topic.tag', searchTaxonomyKey);
-
-      return this.http.get(this.RMMAPIURL + 'records?',
-        {
-          params: params
-
-        });
-    }
-  }
-
     /**
-     * Behavior subject to remotely set the search value. 
+     * Returns an Observable for the HTTP GET request for the JSON resource.
+     * @return {string[]} The Observable for the HTTP request.
      */
-    private _remoteQueryValue : BehaviorSubject<object> = new BehaviorSubject<object>({queryString: '', searchTaxonomyKey: '', queryAdvSearch: 'yes'});
-    public _watchQueryValue(subscriber){
-        this._remoteQueryValue.subscribe(subscriber);
+    searchPhraseTest(searchValue: string, searchTaxonomyKey: string, queryAdvSearch: string): Observable<any> {
+        if ((queryAdvSearch === 'yes' && (!(_.includes(searchValue, 'searchphrase'))))) {
+        return this.http.get(this.RMMAPIURL + 'records?' + searchValue);
+        } else {
+        let params = new HttpParams();
+        params.set('searchphrase', searchValue);
+        params.set('topic.tag', searchTaxonomyKey);
+
+        return this.http.get(this.RMMAPIURL + 'records?',
+            {
+            params: params
+
+            });
+        }
     }
 
-    public setQueryValue(queryString: string = "", searchTaxonomyKey: string = '', queryAdvSearch: string = 'yes') {
-        if(queryString == null) queryString = "";
-        let lq = queryString.replace(/searchphrase=/g, '')
-        this._remoteQueryValue.next({queryString: lq, searchTaxonomyKey: searchTaxonomyKey, queryAdvSearch: queryAdvSearch});
-    }
-
-    /**
-     * Start search
-     */
-    public search(searchValue: string, url?: string) : void {
-        let params: NavigationExtras = {
-            queryParams: {
-                'q': searchValue
-            }
-        };
-
-        if(url == '/search'){
-            this.router.routeReuseStrategy.shouldReuseRoute = () => false;
-            this.router.onSameUrlNavigation = 'reload';
+        /**
+         * Behavior subject to remotely set the search value. 
+         */
+        private _remoteQueryValue : BehaviorSubject<object> = new BehaviorSubject<object>({queryString: '', searchTaxonomyKey: '', queryAdvSearch: 'yes'});
+        public _watchQueryValue(subscriber){
+            this._remoteQueryValue.subscribe(subscriber);
         }
 
-        this.router.navigate(['/search'], params);
-    }
+        public setQueryValue(queryString: string = "", searchTaxonomyKey: string = '', queryAdvSearch: string = 'yes') {
+            if(queryString == null) queryString = "";
+            let lq = queryString.replace(/searchphrase=/g, '')
+            this._remoteQueryValue.next({queryString: lq, searchTaxonomyKey: searchTaxonomyKey, queryAdvSearch: queryAdvSearch});
+        }
+
+        /**
+         * Start search
+         */
+        public search(searchValue: string, url?: string) : void {
+            let params: NavigationExtras = {
+                queryParams: {
+                    'q': searchValue
+                }
+            };
+
+            if(url == '/search'){
+                this.router.routeReuseStrategy.shouldReuseRoute = () => false;
+                this.router.onSameUrlNavigation = 'reload';
+            }
+
+            this.router.navigate(['/search'], params);
+        }
 }
